move: useLoginRedirection from useAuthentication
