+++ src/hooks/useAuthentication.ts
@@ -125,19 +125,3 @@ export const useDetectAuthStateChanged = () => {
 
   return { isInitialized };
 };
-
-export const useLoginRedirection = () => {
-  const history = useHistory();
-  const { isLoggedIn } = useAuthentication();
-
-  useEffect(() => {
-    if (!isLoggedIn) {
-      const searchParams = new URLSearchParams();
-      searchParams.append('from', `${history.location.pathname}${history.location.search}`);
-      // TODO: feedback message
-      history.replace({ pathname: PAGE_PATH.login, search: searchParams.toString() });
-    }
-    // NOTE: run when isLoggedIn changed
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [isLoggedIn]);
-};
